requirements.txt
requests~=2.31.0
PyAutoGUI~=0.9.54
Pillow~=10.1.0
pip~=23.3.2
wheel~=0.42.0
docutils~=0.20.1
tornado~=6.4
filelock~=3.13.1
Kivy~=2.3.0rc2
Pygments~=2.17.2
pygame~=2.1.1
numpy~=1.26.2
certifi~=2023.11.17
setuptools~=69.0.3
pywin32~=306
typing_extensions~=4.9.0
idna~=3.6
urllib3~=1.26.18
cffi~=1.16.0
attrs~=23.2.0
regex~=2023.12.25
validators~=0.22.0
future~=0.18.3
ipython~=8.20.0
pyinstaller~=6.3.0
python-dotenv~=1.0.0
openmeteo_requests~=1.1.0
pandas~=2.1.4
geopy~=2.4.1